* moved SymfonyToLegacyHelper to /src/Domain/Services/ folder * fixed services in services.yaml
