+++ config-symfony5.4/services.yaml
@@ -18,8 +18,8 @@ services:
 
     # makes classes in src/ available to be used as services
     # this creates a service per class whose id is the fully-qualified class name
-    App\:
-        resource: '../src/{Command,Controller}'
+    App\Domain\:
+        resource: '../src/Domain/{Command,Controller,Services}'
         public: true
 
     # add more service definitions when explicit configuration is needed
@@ -29,5 +29,3 @@ services:
     App\Domain\API\APIHelper:
         arguments:
             $projectDir: '%kernel.project_dir%'
-
-    App\Domain\Helper\SymfonyToLegacyHelper:
